framerates #1138 (#326)
* framerates #1138


https://github.com/owncast/owncast/issues/1138#
This is a technically correct and low to no risk making the low end default 25.

now has 25 default, 30, 50, 60 and outrageous. This is a good balance as the standard in NA and Japan and such and such is still 30/60 and EU etc is 25/50. So I've just piggybacked the same description for 50 and 60 options and tacked an EU and NA on them.

Pardon my ignorance but I cannot find where the heck the default visual stream options name only is being generated. I've tried reinstalling owncast on top my edited admin no dice.

tested with all variants individual and stacked with no issues.

* Update config-constants.tsx

per discussion reintroduced the 24fps default thus removing any paradox with the core code text visual naming
removed NA , EU references

re-tested