Use bcrypt hash comparisons for streaming key (#2495)
The streaming key comparison was previously using the `==` operator
between a plaintext version of the streaming key and the one configured
and stored.  This comparison exposes the possibility of a time-based
attack for discovering the streaming key (see issue #2489).

Project contributor @MFTabriz noted that a better solution would be to
use bcrypt and compare hashes.  This commit switches the comparison over
to it.