Migrate to hashed password/key store (#2489)
In order to improve Owncast security, the admin password and stream keys
are no longer being stored in plaintext, instead using a standard
cryptographic hashing process to store them securely.

This commit not only moves the admin password to storage as a hash, but
also splits the StreamKey model into StreamKeyPlainText and
StreamKeyHashed to make it clear when, where, and how a stream key is
being used in its plaintext and hashed forms.  Attempting to store a
plaintext key to the database will automatically hash it.

Since it is no longer possible to send the plaintext admin password and
streaming keys to the web UI, they have been removed from the relevant
HTTP response objects.