Change to service instance approach instead of global module state
At a high level, this branch serves to change many of the modules from
a global state strategy to a service-based strategy. One of the main
changes is in the ActivityPub layer, which includes the persistence
layer.

To facilitate this change, it was necessary to create a high level
abstraction which contains the instances of the sub-services. This
high-level abstraction is located in `~/app` and is what the
`main()` function now instantiates. This `App` struct contains the
initialization logic for the sub-services, as well as the route
handler initialization logic.

Another change was splitting up the route handler initialization
logic, as many of the routes are coupled to the app sub-services
and controllers. They are coupled in that their initializer
methods must be passed a refernce to one of the services. This is
due to the fact that the exported module-level functions have now
been converted to methods of the service instances.